# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E.P.i.C. Youth Ministry Magazine API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Magazine {
  id: number;
  title: string;
  /** e.g. "Summer 2025", "Vol. 1 Issue 3" */
  edition: string;
  publishedAt: string;
  coverImageUrl: string;
  /** @nullable */
  pdfUrl?: string | null;
  /** @nullable */
  description?: string | null;
  featured?: boolean;
}

export interface MagazineInput {
  title: string;
  edition: string;
  publishedAt: string;
  coverImageUrl: string;
  /** @nullable */
  pdfUrl?: string | null;
  /** @nullable */
  description?: string | null;
  featured?: boolean;
}

export interface TeamMember {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  department?: string | null;
  /**
     * JSON string of social links
     * @nullable
     */
  socialLinks?: string | null;
}

export interface Story {
  id: number;
  title: string;
  excerpt: string;
  /** @nullable */
  content?: string | null;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  author?: string | null;
  /** @nullable */
  publishedAt?: string | null;
  featured?: boolean;
}

export interface Event {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  date: string;
  location: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  registrationUrl?: string | null;
}

export type OpportunityCategory = typeof OpportunityCategory[keyof typeof OpportunityCategory];


export const OpportunityCategory = {
  volunteer: 'volunteer',
  job: 'job',
  scholarship: 'scholarship',
  internship: 'internship',
  resource: 'resource',
} as const;

export interface Opportunity {
  id: number;
  title: string;
  category: OpportunityCategory;
  description: string;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * For scholarships/jobs, the value or salary
     * @nullable
     */
  amount?: string | null;
}

export interface PrayerRequestInput {
  name: string;
  /** @nullable */
  email?: string | null;
  request: string;
  anonymous?: boolean;
}

export interface PrayerRequest {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  request: string;
  anonymous?: boolean;
  submittedAt: string;
}

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  subject?: string | null;
  message: string;
}

export interface Contact {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  subject?: string | null;
  message: string;
  submittedAt: string;
}

export interface SiteStats {
  totalMagazines: number;
  totalTeamMembers: number;
  totalStories: number;
  totalOpportunities: number;
  totalEvents: number;
  youthReached?: number;
  volunteerHours?: number;
  communitiesServed?: number;
}

export type ListMagazinesParams = {
/**
 * Number of issues to return
 */
limit?: number;
/**
 * Offset for pagination
 */
offset?: number;
};

export type ListStoriesParams = {
/**
 * Filter to featured stories only
 */
featured?: boolean;
limit?: number;
};

export type ListEventsParams = {
/**
 * Filter to upcoming events only
 */
upcoming?: boolean;
limit?: number;
};

export type ListOpportunitiesParams = {
/**
 * Filter by opportunity category
 */
category?: ListOpportunitiesCategory;
};

export type ListOpportunitiesCategory = typeof ListOpportunitiesCategory[keyof typeof ListOpportunitiesCategory];


export const ListOpportunitiesCategory = {
  volunteer: 'volunteer',
  job: 'job',
  scholarship: 'scholarship',
  internship: 'internship',
  resource: 'resource',
} as const;