# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E.P.i.C. Youth Ministry Magazine API
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  name: string;
  email: string;
  /** @nullable */
  subject?: string | null;
  message: string;
}