# lib/api-zod/src/generated/types/opportunity.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E.P.i.C. Youth Ministry Magazine API
 * OpenAPI spec version: 0.1.0
 */
import type { OpportunityCategory } from './opportunityCategory';

export interface Opportunity {
  id: number;
  title: string;
  category: OpportunityCategory;
  description: string;
  /** @nullable */
  organization?: string | null;
  /** @nullable */
  deadline?: Date | null;
  /** @nullable */
  applyUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /**
     * For scholarships/jobs, the value or salary
     * @nullable
     */
  amount?: string | null;
}