# lib/api-zod/src/generated/types/prayerRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E.P.i.C. Youth Ministry Magazine API
 * OpenAPI spec version: 0.1.0
 */

export interface PrayerRequest {
  id: number;
  name: string;
  /** @nullable */
  email?: string | null;
  request: string;
  anonymous?: boolean;
  submittedAt: Date;
}