# lib/api-zod/src/generated/types/siteStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E.P.i.C. Youth Ministry Magazine API
 * OpenAPI spec version: 0.1.0
 */

export interface SiteStats {
  totalMagazines: number;
  totalTeamMembers: number;
  totalStories: number;
  totalOpportunities: number;
  totalEvents: number;
  youthReached?: number;
  volunteerHours?: number;
  communitiesServed?: number;
}