# lib/api-zod/src/generated/types/story.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E.P.i.C. Youth Ministry Magazine API
 * OpenAPI spec version: 0.1.0
 */

export interface Story {
  id: number;
  title: string;
  excerpt: string;
  /** @nullable */
  content?: string | null;
  category: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  author?: string | null;
  /** @nullable */
  publishedAt?: Date | null;
  featured?: boolean;
}