# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * E.P.i.C. Youth Ministry Magazine API
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMember {
  id: number;
  name: string;
  role: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  department?: string | null;
  /**
     * JSON string of social links
     * @nullable
     */
  socialLinks?: string | null;
}